+++ agb/src/sync/locks.rs
@@ -142,6 +142,7 @@ pub struct InitOnce<T> {
     is_initialized: Static<bool>,
     value: UnsafeCell<MaybeUninit<T>>,
 }
+
 impl<T> InitOnce<T> {
     /// Creates a new uninitialized object.
     #[must_use]
@@ -201,6 +202,13 @@ impl<T> InitOnce<T> {
         }
     }
 }
+
+impl<T> Default for InitOnce<T> {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 impl<T> Drop for InitOnce<T> {
     fn drop(&mut self) {
         if self.is_initialized.read() {